+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Per school email
 - Per school participant export
 - Send only one email par orga
+- Send email communicating time slots
 
 ## Version 3.0.0 - 2022-12-02 - Interludes 2023 Saclay